cdk-map-tagger/MapTaggingAutomatorService/requirements.txt
-e .

aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-iam

aws_cdk_test_synth